mint-arena: Allow client to disable gib effect completely
In vanilla Quake 3, com_blood 0 disables blood and gib effect. It affects
both game and cgame, so in multiplayer, clients do not have full control
of whether gib effect is shown. If com_blood 1 on server and 0 on client,
corpses just vanish.

This commit makes game send clients the needed information for both gib
and non-gib death effects. com_blood 0 or cg_gibs 0 on client now completely
disables the gib effect. The corpse will be shown instead.

The down side to this is corpses are shootable until gibbed even if client
does not show gib effect. This is mainly noticeable with com_blood 1 and
cg_gibs 0 because blood hit effect is shown until gibbed. Whereas in Q3,
if com_blood is 0 on server and client, the corpses are always shootable.

This changes the gameplay in Team Arena with com_blood 0 a little.
In Team Arena if a player is killed while they are holding the Kamikaze,
players can gib the corpse to stop the Kamikaze from exploding.
Vanilla Team Arena with com_blood 0 does not allow corpse to be gibbed,
so it always explodes. This commit always allows corpse to be gibbed.

Relatively minor issues and has the benefit of working the same regardless
of server settings.

----

This break game <-> cgame and demo compatibility, so GAME_VERSION was changed.

cg_gibs was changed to disable gib effect and removing body. Before it only
disabling spawning most of the gibs (for performance).

Unsetting EF_KAMIKAZE in G_StartKamikaze/GibEntity fixes Kamikaze skull
orbiting corpse after gibbed if client has gib effect disabled.