mint-arena: Disable overbounce bug by default
Jumping in place sometimes causes Z-velocity to flip direction
when they hit the ground, causing player to "jump" again.

This is a known Q3 bug, but like the strafing bug it's considers
a gameplay feature by some. However, the overbounce bug looks
really dumb and makes no sense so disable it by default.

Require server to set pmove_overbounce 1 to enable this behavior.