mint-arena: Place player sprites over tag_torso, instead of entity origin
Use tag_torso on player's legs model as origin for floating
award/talk/etc sprites. This causes sprites to move with the
model when jumping and stuff, instead of being attacked to an
invisible object (the entity).

Also don't draw connection and talk sprites not-in-mirrors
when in first person. (I broke it.)