mint-arena: Fix bot activating button and not entering area
If bot is blocked by something for too long then he will not enter the area he opened.
In other words he pushed the button and the door opened, but he did not enter through the door (which looks weird).

This bugfix does not stop the bot from choosing a new goal, such as hunting a player, which may cause the bot not enter the opened area.
So this bugfix only attracts the bot more to enter enabled areas if the goal didn't change while pressing buttons etc.