mint-arena: AltUI: Add resolution and geometric list
Build list of resolusions at start up, complete with commands for custom modes.
The initial resolution 'pair index' is saved for initial pair for menu item.

Each time the system menu is opened it uses the initial pair for resolution,
rather than the one last selected in UI. Though, this wouldn't be an issue if
the value was not set until apply button was clicked -- currently there is no
apply button.

Geometric Detail now has a list of command to run, but does not detect what
the initial pair should be. Graphics Settings has this issue too.
Need to allow specifying a item function for detecting which pair to use?