mint-arena: Remove unused armor and holdable sphere model loading
Shard armor and medkit holdable list a model to be used as a sphere.
They are not rendered because the "accompanying ring/sphere" is only
used if item is IT_POWERUP or IT_HEALTH. So don't attempt to load them.

The shard_sphere.md3 exists (shape matches shard icon :-), but is not
meant to be used (it's texture is opaque, no shader).
shard.md3 has a effects surrounding built in anyway.
The medkit_sphere.md3 does not exist.