mint-arena: Don't send GTS_*_CAPTURE to player who captured
A local sound is played for the capture award. The global sound is the
same sound effect and should not be played twice. So only send global sound
to everyone except the player (and their fellow local players) who captured.

If a local player is spectate following a non-local player the double sound issue
will still happen though.