mint-arena: Reduce MAX_CHARACTERISTICS to save memory
In Quake 3 it's in the engine, so by having it higher than needed mods
could add more characteristics in the Game VM. In mint-arena it's in the
Game VM, so mods can revert this or increase it if needed.

I increased the number of characteristic arrays by about 3 times in the
last commit. This is an attempt about getting closer to original memory
usage.