mint-arena: Improve edit field cursor drawing
Make cursor be the same width as the character it's selecting (for
proportional width fonts).

If in overstrike mode and cursor is at end of buffer a new character will be
inserted so use low line cursor instead.

Draw cursor behind selected character and invert color of character when the
overstrike cursor is visible so that the character can be read when full
width block cursor is visible.

Added a hack for Team Arena fonts to convert vertical pipe into a full width
block for overstrike cursor.