mint-arena: Make Travis CI only run Coverity on a separate branch
Sometimes builds fail because Coverity refused to run because it's
rate limited (few times a week). Also, Coverity needs to only be run
once which can only be done using a separate branch.

Not really ideal, but ease of use of pushing a branch makes it better
than manually running Coverity scan tool and uploading a tar.