mint-arena: Allow disabling TrueType / Team Arena dat fonts
This makes it easier to compare TrueType and Q3 bitmap fonts without
changing code or moving files (which might not be possible if fonts
are in a pk3).

Added bitmap fallbacks for Team Arena UI and HUD. It's now easier than
ever to play Team Arena using gfx/2d/bigchars.tga as the only font!