mint-arena: Give bot more time to get item after activating entity
When bot goes for a nearby item, there is a limit to how long the bot
will spend trying to get the item. However, the bot does not calculate
how long that will take if entities must be activated to get the item.
After activating the entities the bot switches back to seeking nearby
goal and may immediately pick something new because the bot has taken
too long to pick it up.

Work around this issue by giving the bot more time to pick up the nearby
goal if it successfully activate the entities.

This issue is present on q3dm7 when going for the heavy armor.