mint-arena: Allow persistant powers without team to be picked up
If Team Arena persistant power up item's spawnflags doesn't have either
red or blue team bit set, let everyone pick it up. This is the behavior
in vanilla Team Arena.

mint-arena also allows setting both red and blue team bits to allow
everyone to pick it up but in Team Arena it causes only red team to
be able to pick it up. So it's not useful for making maps for both
Spearmint and vanilla Team Arena.