mint-arena: Add view weapon fov to CGame API
This solves many issues that were encounted when putting view
weapon in a separate scene with it's own field of view. Such as
lights in separate scenes and lightning trail not appearing in
mirrors.

While it probably would of been possible to work around the issues
of using separate scenes, I really don't want to have to deal with
it for each game that gets ported.