mint-arena: Always display non-responsive servers
It's annoying when q3_ui says 3 of 3 Arena Servers but only 2 are
in the list. So add non-responsive servers to the list for Internet
and local, not just favorites.

This is a bigger problem for Spearmint than ioq3 because there are
few servers which makes it easy for users to count servers and see
that the number is wrong.