mint-arena: Add cg_crosshairHealth to q3_ui menu
Requires changing Q3's shaders to use rgbGen exactVertex
instead of rgbGen identity (which was colorized in pre-
release versions of Quake 3).

The crosshair is colorized in the menu to give a better
understanding of what "crosshair health" means as well
as a quick way to see if shaders have been fixed or not.