mint-arena: Fix playing demos on macOS by opening demo via finder
macOS send a file drop event when opening a demo file associated with
Spearmint. This results in there being no launch arguments so CGame
queues the logo cinematic to play later. However, the file drop is
processed first causing map to load for demo and then switches to
playing the logo cinematic.