mint-arena: Fix using clientnum as playernum in G_TeamCommand
If you start a team deathmatch game with a splitscreen player
and a bot (added in that order), team server commands intended
for bot (client 1) got sent to the splitscreen player (player 1)
instead.

It caused duplicate team chat messages for the splitscreen player
and the bot did not receive it.