mint-arena: Fix lightning trail being fired from corpses
The lightning trail is fired from the real non-predicted entity for
the local client so it can show the server location. The real client
entity was always used to check if a player entity is firing the
lighting gun. This potentially causes corpses and players on single
player podiums with lightning gun or missing weapon model to render
the lightning trail if the real (living) client is firing the
lightning gun.

This is an issue for corpses as they have WP_NONE and thus always
missing the weapon model. It's not an issue for podium players as
it's not possible to fire during intermission.

Note: This was committed to ioquake3, which doesn't draw lightning if
missing weapon model, as
"Fix lightning gun handling for corpses and single player podiums".